Renamed pdo-repository to model
